backend/pyproject.toml
[project]
name = "tennis_web_backend"
version = "0.1.0"
description = "Backend for a Tennis web application"
authors = [{ name = "Valentine", email = "vchidubem9@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "anyio==4.9.0",
    "async-timeout==5.0.1",
    "bidict==0.23.1",
    "certifi==2025.4.26",
    "click==8.2.1",
    "dnspython==2.7.0",
    "email_validator==2.2.0",
    "fastapi==0.115.12",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "Jinja2==3.1.6",
    "MarkupSafe==3.0.2",
    "pydantic==2.11.3",
    "pydantic_core==2.33.1",
    "python-dotenv==1.1.0",
    "python-engineio==4.11.2",
    "python-multipart==0.0.20",
    "python-socketio==5.13.0",
    "PyYAML==6.0.2",
    "redis==6.4.0",
    "simple-websocket==1.1.0",
    "sniffio==1.3.1",
    "starlette==0.46.2",
    "typing_extensions==4.13.1",
    "uvicorn==0.34.2",
    "uvloop==0.21.0",
    "watchfiles==1.0.5",
    "websockets==15.0.1",
    "wsproto==1.2.0",
    "GameEngine @ git+ssh://git@github.com/Theocode12/GameEngine.git@v0.3.0",
]

[project.optional-dependencies]
dev = [
    "annotated-types==0.7.0",
    "fastapi-cli==0.0.7",
    "markdown-it-py==3.0.0",
    "mdurl==0.1.2",
    "mypy==1.15.0",
    "mypy_extensions==1.1.0",
    "Pygments==2.19.1",
    "rich==14.0.0",
    "rich-toolkit==0.14.6",
    "ruff==0.11.5",
    "shellingham==1.5.4",
    "StrEnum==0.4.15",
    "typer==0.16.0",
    "typing-inspection==0.4.0",
]

[tool.setuptools]
packages = []

[tool.ruff]
exclude = [".git", ".ruff_cache", ".venv"]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "C4", "B", "RUF"]
ignore = ["E501", "E731", "B008", "RUF013"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
known-first-party = ["gameengine"]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.ruff.lint.per-file-ignores]
"test/**/*" = ["S101", "ANN"] # ignore all assertion rules in tests

[tool.mypy]
# Use type checking for files in app
files = ["app/**/*.py"]
# Check dynamically typed code
check_untyped_defs = true
# Raise errors for any errors
strict = true
# Type checking with debug
show_error_codes = true
show_column_numbers = true
disable_error_code = ["import-untyped"]

[[tool.mypy.overrides]]
module = ["*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
